dcp-instance-manager/requirements.txt
flask
flask_cors
python-dotenv
dnspython
firebase-admin
requests